pyproject.toml
[project]
name = "spec-to-pr"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.50.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[project.scripts]
spec-to-pr = "spec_to_pr.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/spec_to_pr"]

[tool.pytest.ini_options]
testpaths = ["tests"]
